fix: update from igor to igor2
igor no longer works with numpy as of numpy 1.20. If trying to
import SciFiReaders with igor, we get the following error:

"AttributeError: module 'numpy' has no attribute 'complex'."

indicating the issue. Because of this deprecation issue, AFM-Analysis
have forked the original wking/igor, and updated it into a new
repository called igor2.

This CL updates SciFiReaders to use igor2 instead of igor. The igor
unit tests have been tested and pass. The changes here correspond
exclusively to changing text from 'igor' to 'igor2' for package
purposes and user prompts (we do not change any class/variable
names).

Note: for 2 files, the diff will show the addition of a \newline
at the end of the file. This is unix policy and done by default
by my editor.