fix: force igor2 version to de-risk issue
Thinking about it, the 'real' issue was that our igor2 version did not
require a strict version. If we set igor2 to the current version (with
these changes), we should minimize the likelihood of this issue propping
up again.

Note, however, that any improvements of igor2 will be lost. A more
'proper' solution would likely be to mix a more general pyproject.toml
with a 'lock' file defining the versions that worked for the developers
of SciFiReaders.