feat: update pifm.py for 'new' reader rules
The new pycroscopy has switched from Translators to Readers. We no
longer save files in USID or NSID format in order to translate, but
rather translate while reading from the file. The user can choose to
save a given sidpy dataset into USID or NSID format after-the-fact, via
those libraries.

Here, we update the pifm module to match this new standard. As such, we
remove the pyNSID dependencies and the write/create methods which use them.